Reorg testing code.
