+++ tests/simulator/contract_interfaces/appchain_lifecycle_manager.rs
