+++ tests/simulator/contract_interfaces/mod.rs
@@ -1,0 +1,7 @@
+pub mod appchain_lifecycle_manager;
+pub mod appchain_owner_actions;
+pub mod registry_roles;
+pub mod registry_settings;
+pub mod registry_viewer;
+pub mod sudo_actions;
+pub mod voter_actions;
