+++ tests/simulator/contract_interfaces/registry_roles.rs
