+++ tests/simulator/contract_interfaces/registry_settings.rs
