+++ tests/simulator/contract_interfaces/registry_viewer.rs
